Add files via upload
A Random OTP (One-Time Password) Generator is a web-based tool that generates a random numeric or alphanumeric code that can be used for secure user verification. It’s commonly used in login, signup, or password reset systems.

This project is usually built using HTML, CSS, and JavaScript, and it's ideal for beginners to understand random number generation, and basic frontend logic.